Update int to double in CapCommand

+++ src/main/java/planner/command/CapCommand.java
@@ -167,7 +167,7 @@ public class CapCommand extends ModuleCommand {
             if (!detailedMap.containsKey(userInfo[0].toUpperCase())) {
                 throw new ModNotFoundException();
             }
-            int mcTemp = detailedMap.get(userInfo[0].toUpperCase()).getModuleCredit();
+            double mcTemp = detailedMap.get(userInfo[0].toUpperCase()).getModuleCredit();
             if (!detailedMap.get(userInfo[0].toUpperCase()).getAttributes().isSu()
                 || letterGradeToCap(userInfo[1].toUpperCase()) != 0.00) {
                 mcCount += mcTemp;
